add netfuse client and server and have cpu use it
cpu currently uses the 9p protocol. The server for 9p is
provided by an external program. We are experimenting
with hugelgupf's excellent p9 package but it can't resolve
the problems that attend to 9p.

It makes sense to provide it internally because, since
the remote 9p requests are funneled through cpu, we would
avoid the extra hop. Note that in Plan 9 cpu that's not how it
works but for a lot of reasons we use ssh port remapping
to move requests from the remote machine to the machine
we cpu from. Long story.

In an ideal world, the Linux 9p maintainers would
realize that 9p is a resource sharing protocol, not
a file protocol, and would not make gratuitous changes
that break it in simple ways, making it no longer
work with extant 9p servers. We've had a Go 9p server
in Harvey working for 4 years now; Linux can't use it
because Linux broke how reads work in the 9p client.

In an ideal world, we'd have a better protocol than 9p
which does not have the serialization and other
performance-limiting aspects of 9p. Let's face it, 9p
is slow, and it's for reasons that are organic to the
protocol. They're not fixable and they get worse as
networks get faster.

We don't live in an ideal world.

This PR has two pieces:
1. the creation of a netfuse package. netfuse takes the
server from jacobsa/fuse and puts a Go RPC in the middle.
Note that Go RPC by default uses gob so it's not too bad.
The result is that a program (like cpu) can run a FUSE
server and relay the requests to an RPC server. Then, a program
(like cpu) can actually perform the operations and return
the result. Note we had to fork jacobsa/fuse to u-root/fuse
for now (we hope to go back) because jacobsa/fuse does not
pass open flags to the FUSE server and that's just not tenable.
jacobsa seems kinda busy and we really need this change.

2. cpu can now use netfuse. Note that the 9p code is still there
and still working; you can use 9p if you wish via -use9p.
There's no compelling reason to remove it, since -use9p does
not actually embed 9p support in cpu.

This works just fine. FUSE has proven to be a better
match to the Linux VFS.

For testing, add fuse to the docker containers.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>