cmds/core/msr: cleanup
Don't panic on empty commands [!]

Add some synonyms for the simple usage:
if os.Args[0] is wrmsr or rdmsr, act like the standard tool

if flags.Args[0] is wrmsr, wr, rdmsr, rd, act like the standard tool

make sure panics push an error, not a string.

panics are still a perfectly adequate way to handle parser errors,
per Rob and Russ.

forth will now handle all panics, not just os panics. Not sure
why it was done the way it was done, but that was a long time ago.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>