Speed up dd
A common usage of dd is to have identical ibs and obs, i.e.
bs= is used.

In that case, we can use io.CopyBuffer.

A problem with this approach is that we won't really have the progress bar.

Note that the slow case is 10 GiB/s, which is not bad.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>